% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{annotate_cell_type}
\alias{annotate_cell_type}
\title{Annotate unified cell_type names}
\usage{
annotate_cell_type(result_table, method)
}
\description{
(map the cell_types of the different methods to a common name)
}
