% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\docType{data}
\name{cell_type_list}
\alias{cell_type_list}
\title{List with controlled cell-type vocabulary}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74 rows and 3 columns.
}
\usage{
cell_type_list
}
\description{
List with controlled cell-type vocabulary
}
\keyword{datasets}
