% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\docType{data}
\name{cell_type_mapping}
\alias{cell_type_mapping}
\title{a mapping of the various cell_types between the different methods to a common standard}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 8 columns.}
\usage{
cell_type_mapping
}
\description{
a mapping of the various cell_types between the different methods to a common standard
}
\keyword{datasets}
