% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_datasets.R
\docType{data}
\name{celltype2dataset_mapping}
\alias{celltype2dataset_mapping}
\title{a mapping of the various cell_types between the different methods to a common standard}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 6 columns.}
\usage{
celltype2dataset_mapping
}
\description{
a mapping of the various cell_types between the different methods to a common standard
}
\keyword{datasets}
