% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_deconvolution_methods.R
\name{convert_human_mouse_genes}
\alias{convert_human_mouse_genes}
\title{This function converts the mouse gene symbols into corresponding human ones, and vice versa.}
\usage{
<<<<<<< HEAD:man/convert_human_mouse_genes.Rd
convert_human_mouse_genes(
  gene_expression_matrix,
  mirror = "www",
  other_annot = TRUE,
  convert_to = c("human", "mouse")
=======
mouse_genes_to_human(
  gene_expression_matrix,
  mirror = "www",
  other_annot = TRUE
>>>>>>> a7e4ee9993aa94f268e862263eaf226a251514f9:man/mouse_genes_to_human.Rd
)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples.
Gene symbols must be the rownames of the matrix.}

\item{mirror}{the ensembl mirror to use. Possible choices are 'www' (default),
'uswest', 'useast', 'asia'}

\item{other_annot}{boolean, wether to run the other conversion method (might be memory intensive)}

\item{convert_to}{one of 'human' or 'mouse'. Specifies the organism of the orthologs to look for}
}
\value{
the same matrix, with the counts for the corresponding human genes.
This matrix can directly be used with the immunedeconv methods. A message
will display the ratio of original genes which were converted.
}
\description{
This function relies on the `biomaRt`` package and connects to the ENSEMBL repository
to retrieve the gene symbols. If ENSEMBL cannot be reached, another solution will be
used. Since it is memory intensive, users can choose not to run it.
}
