% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE.R
\name{create_base_compendium}
\alias{create_base_compendium}
\title{Souce code to create the compendium used kin the BASE algorithm, containing
up- and down-regulated weight sets that specify the
This code is adapted from Varn et al., DOI: 10.1158/0008-5472.CAN-16-2490}
\usage{
create_base_compendium(signature_matrix)
}
\arguments{
\item{signature_matrix:}{numeric matrix; The signature matrix from which the compendium will be built.
Must contain genes on rows and cell on columns}
}
\description{
Souce code to create the compendium used kin the BASE algorithm, containing
up- and down-regulated weight sets that specify the
This code is adapted from Varn et al., DOI: 10.1158/0008-5472.CAN-16-2490
}
