% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_deconvolution_methods.R
\name{custom_deconvolution}
\alias{custom_deconvolution}
\title{Collection of deconvolution methods that allow custom signature matrices.}
\description{
Collection of deconvolution methods that allow custom signature matrices.
}
