% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_racle}
\alias{dataset_racle}
\title{Example RNA-seq dataset from the EPIC publication.}
\format{
an environment with two objects:
(1) \code{expr_mat}: gene expression matrix with gene_symbols as rownames and sample identifiers as colnames.
(2) \code{ref}: FACS measurements in a data.frame with three columns: sample, cell_type, true_fraction
}
\source{
Racle et al. (2017), eLIFE, https://doi.org/10.7554/eLife.26476.029
}
\usage{
dataset_racle
}
\description{
A dataset of four RNA-seq samples from patients with metastatic melanoma and associated
'gold standard' immune cell estimates profiled with fluorescence activated cell sorting (FACS)
}
\keyword{datasets}
