% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\docType{methods}
\name{deconvolute}
\alias{deconvolute}
\alias{deconvolute.default}
\alias{deconvolute}
\alias{deconvolute,eSet,character-method}
\alias{deconvolute,matrix,character-method}
\title{Perform an immune cell deconvolution on a dataset.}
\usage{
deconvolute.default(gene_expression, method = deconvolution_methods, ...)

deconvolute(gene_expression, method = deconvolution_methods,
  column = "gene_symbol")

\S4method{deconvolute}{eSet,character}(gene_expression,
  method = deconvolution_methods, column = "gene_symbol")

\S4method{deconvolute}{matrix,character}(gene_expression, method)
}
\arguments{
\item{gene_expression}{A number.}

\item{method}{a string specifying the method. Supported methods are \code{xcell}, \code{...}}

\item{...}{arguments passed to the respective method}

\item{eset}{Expression set}

\item{col}{column name of the \code{fData} column that contains HGNC gene symbols.}
}
\value{
\code{data.frame} with \code{cell_type} as first column and a column with the
calculated cell fractions for each sample.
}
\description{
Perform an immune cell deconvolution on a dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{gene_expression = eSet,method = character}: \code{eset} is a \code{Biobase::ExpressionSet}.
\code{fData} contains a column with HGNC gene symbols (specify column name)

\item \code{gene_expression = matrix,method = character}: \code{matrix} is matrix with HGNC gene symbols as row names.
}}

\examples{
deconvolute(gene_expression_matrix, "xcell")

}
