% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolute_abis}
\alias{deconvolute_abis}
\title{Deconvolute using ABIS.}
\usage{
deconvolute_abis(gene_expression_matrix, arrays = FALSE, ...)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples. Data
must be TPM-normalized, since the matrix accounts for mRNA bias}

\item{arrays}{Set to TRUE if working with Microarray data instead of RNA-seq.
A different signature matrix will be used.}

\item{...}{additional arguments}
}
\description{
Deconvolute using ABIS.
}
