% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abis.R
\name{deconvolute_abis_default}
\alias{deconvolute_abis_default}
\title{Use ABIS to deconvolute a gene expression matrix.}
\usage{
deconvolute_abis_default(expression, array = FALSE)
}
\arguments{
\item{expression}{numeric matrix; genes as rows and patients as columns.
Must be TP normalized}

\item{array:}{logical; specifies whether the expression data are from
microarrays (TRUE) or RNAseq (FALSE)}
}
\description{
Source code adapted from https://github.com/giannimonaco/ABIS -
server.R of the Shiny application
}
