% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_deconvolution_methods.R
\name{deconvolute_base_algorithm}
\alias{deconvolute_base_algorithm}
\title{Deconvolute using BASE}
\usage{
deconvolute_base_algorithm(
  gene.expression.matrix,
  n.permutations = 100,
  log10 = TRUE,
  combine_cells = TRUE,
  ...
)
}
\arguments{
\item{gene.expression.matrix}{a m x n matrix with m genes and n samples. Data
should be TPM normalized and log10 scaled.}

\item{n.permutations}{the number of permutations of each sample expression
to generate. These are used to normalize the results.}

\item{log10}{logical. if TRUE, log10 transforms the expression matrix.}

\item{combine_cells}{logical. This method estimates several cell types (~180).
If TRUE (default), these are combined into 19 major immune-stromal cell types.}
}
\description{
Cell types obtained with this method can be reduced with the
map_result_to_celltypes function, using 'DCQ' as the method
}
