% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_deconvolution_methods.R
\name{deconvolute_base_custom}
\alias{deconvolute_base_custom}
\title{Deconvolute using BASE and a custom signature matrix}
\usage{
deconvolute_base_custom(
  gene_expression_matrix,
  signature_matrix,
  n_permutations = 100,
  log10 = TRUE
)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples. Data
should be TPM normalized and log10 scaled.}

\item{signature_matrix}{a m x l matrix with m genes and l cell types. Data
should be non normalized, as the normalization wil be done in the construction
of the compendium (internal structure)}

\item{n_permutations}{the number of permutations of each sample expression
to generate. These are used to normalize the results.}

\item{log10}{logical. if TRUE, log10 transforms the expression matrix.}
}
\description{
Deconvolute using BASE and a custom signature matrix
}
