% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolute_consensus_tme}
\alias{deconvolute_consensus_tme}
\title{Deconvolute using ConsensusTME.}
\usage{
deconvolute_consensus_tme(
  gene_expression_matrix,
  indications = NULL,
  method = "ssgsea",
  ...
)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples}

\item{indications}{a n-vector giving and indication string (e.g. 'brca') for each sample.
The method requires at least 2 samples of a certain cancer type.
Accepted indications are 'kich', 'blca', 'brca', 'cesc', 'gbm', 'hnsc', 'kirp', 'lgg',
'lihc', 'luad', 'lusc', 'prad', 'sarc', 'pcpg', 'paad', 'tgct',
'ucec', 'ov', 'skcm', 'dlbc', 'kirc', 'acc', 'meso', 'thca',
'uvm', 'ucs', 'thym', 'esca', 'stad', 'read', 'coad', 'chol'}

\item{method}{Choose statistical framework to generate the entichment scores.
Default: 'ssgsea'. Available methods: 'ssgsea', 'gsva', 'plage', 'zscore', 'singScore'.
These mirror the parameter options of \code{GSVA::gsva()} with the exception of \code{singScore}
which leverages \code{singscore::multiScore()}}

\item{...}{passed through to the original ConsensusTME function. A native argument takes precedence
over an immunedeconv argument. Documentation can be found at http://consensusTME.org}
}
\description{
Deconvolute using ConsensusTME.
}
