% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_deconvolution_methods.R
\name{deconvolute_consensus_tme_custom}
\alias{deconvolute_consensus_tme_custom}
\title{Deconvolute using ConsesnusTME and a custom signature matrix}
\usage{
deconvolute_consensus_tme_custom(
  gene_expression_matrix,
  signature_genes,
  stat_method = "ssgsea"
)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples. Data
should be TPM normalized and log10 scaled.}

\item{signature_genes}{a list with each element containing genes to represent a cell type. The cell types
should be the names of each element of the list.}

\item{stat_method}{Choose statistical framework to generate the entichment scores.
Default: 'ssgsea'. Available methods: 'ssgsea', 'gsva', 'plage', 'zscore', 'singScore'.
These mirror the parameter options of \code{GSVA::gsva()} with the exception of \code{singScore}
which leverages \code{singscore::multiScore()}}
}
\description{
Deconvolute using ConsesnusTME and a custom signature matrix
}
\note{
ConsensusTME uses tumor-specific consensus built gene signatures. In this case
only the user-provided signature will be used
}
