% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolute_eigengene}
\alias{deconvolute_eigengene}
\title{Deconvolute using eigengene method by Zhang et. al.}
\usage{
deconvolute_eigengene(
  gene_expression_matrix,
  m_signature = eig.sig,
  min_count = 100,
  min_samples_w_count = c(2, 5)
)
}
\arguments{
\item{gene_expression_matrix}{m x n matrix (m genes, n samples) of \strong{counts}
not tpm. rownames must match genes in signature}

\item{m_signature}{data.frame with 2 columns gene, cell_type. Represents
which genes are in signature for which cell type. Default:
signature from \emph{Charoentong et. al. 2017}}

\item{min_count}{minimum number of reads a gene needs to be considered}

\item{min_samples_w_count}{Removes genes where there are less than
\code{min_count_sample[1]} samples with more than \code{min_count_sample[2]} counts}
}
\description{
Deconvolute using eigengene method by Zhang et. al.
}
