% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolute_epic}
\alias{deconvolute_epic}
\title{Deconvolute using EPIC}
\usage{
deconvolute_epic(gene_expression_matrix, tumor, scale_mrna, ...)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples}

\item{tumor}{Set to TRUE if working with tumor data. Will choose the \code{TRef}
signature matrix in that case, \code{BRef} otherwise (through EPIC's \code{reference} parameter)}

\item{scale_mrna}{Set to FALSE to disable correction for cell type-specific differences
in mRNA content (through EPIC's \code{mRNA_cell} parameter)}

\item{...}{passed through to EPIC. A native argument takes precedence
over an immunedeconv argument (e.g. \code{ref} takes precedence over \code{tumor})
See \href{https://rdrr.io/github/GfellerLab/EPIC/man/EPIC.html}{EPIC}}
}
\description{
Deconvolute using EPIC
}
