% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_deconvolution_methods.R
\name{deconvolute_epic_custom}
\alias{deconvolute_epic_custom}
\title{Deconvolute using EPIC and a custom signature matrix.}
\usage{
deconvolute_epic_custom(
  gene_expression_matrix,
  signature_matrix,
  signature_genes,
  genes_var = NULL,
  mrna_quantities = NULL,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples}

\item{signature_matrix}{a m x l matrix with m genes and l cell types. This matrix
should contain the whole set of genes}

\item{signature_genes}{a character vector of the gene names to use as signature
needs to be smaller than the genes in the signature matrix}

\item{genes_var}{(optional) a m x l matrix with m genes and l cell types, with
the variability of each gene expression for each cell type.
This will be used in the optimization}

\item{mrna_quantities}{(optional) A named numeric vector with
the amount of mRNA in arbitrary units for each of the
reference cells and of the other uncharacterized cells.}

\item{...}{passed through to EPIC. A native argument takes precedence
over an immunedeconv argument.
See \href{https://rdrr.io/github/GfellerLab/EPIC/man/EPIC.html}{EPIC}}
}
\description{
Deconvolute using EPIC and a custom signature matrix.
}
