% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_deconvolution_methods.R
\name{deconvolute_immucellai_mouse}
\alias{deconvolute_immucellai_mouse}
\title{Deconvolute using ImmuCellAI-mouse}
\usage{
deconvolute_immucellai_mouse(
  gene_expression_matrix,
  arrays,
  layer2 = FALSE,
  layer3 = FALSE
)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples. Should be TPM normalized}

\item{arrays}{Set to TRUE if microarray data, to FALSE for RNASeq (\code{rnaseq} parameter in ImmuCellAI_mouse)}

\item{layer2}{logical, wether to return the fractions for cell types in layer 2}

\item{layer3}{logical, wether to return the fractions for cell types in layer 3 (if TRUE, this will NOT override layer2)}
}
\description{
Deconvolute using ImmuCellAI-mouse
}
