% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_deconvolution_methods.R
\name{deconvolute_mmcp_counter}
\alias{deconvolute_mmcp_counter}
\title{Deconvolute using mMCP-counter}
\usage{
deconvolute_mmcp_counter(
  gene_expression_matrix,
  log2 = TRUE,
  gene_id = "Gene.Symbol",
  genome = "GCRm39"
)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples. Should be TPM normalized}

\item{log2}{logical. If TRUE, log2 transforms the expression matrix}

\item{gene_id}{specifies the format of the gene IDs. Possible choices are 'Gene.Symbol'
(default), 'ENSEMBL.ID' or 'Probes' (Affymetrix array probes)}

\item{genome}{specifies the mouse genome version to use, GCRm39 (default) or GCRm38}
}
\description{
Deconvolute using mMCP-counter
}
