% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolute_quantiseq}
\alias{deconvolute_quantiseq}
\title{Deconvolute using quanTIseq}
\usage{
deconvolute_quantiseq(gene_expression_matrix, tumor, arrays, scale_mrna,
  ...)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples}

\item{tumor}{Set to TRUE if dealing with a tumor samples. if TRUE, signature genes with
high expressin in tumor samles are removed.}

\item{arrays}{Set to TRUE if working with Microarray data instead of RNA-seq}

\item{scale_mrna}{Set to FALSE to disable correction for cell type-specific differences
in mRNA content}

\item{...}{passed through to original quantiseq method. A native argument takes precedence
over an immunedeconv argument (e.g. \code{mRNAscale} takes precedence over \code{scale_mrna})
See \code{deconvolute_quantiseq.default()}.}
}
\description{
Deconvolute using quanTIseq
}
