% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolute_random}
\alias{deconvolute_random}
\title{Deconvolute using the awseome RANDOM technique}
\usage{
deconvolute_random(gene_expression_matrix)
}
\description{
Here is a good place to add some documentation.
}
