% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_deconvolution_methods.R
\name{deconvolute_seqimmucc_custom}
\alias{deconvolute_seqimmucc_custom}
\title{Deconvolute using seqImmuCC (LLSR regression) and a custom signature matrix}
\usage{
deconvolute_seqimmucc_custom(gene_expression_matrix, signature_matrix, ...)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples. Data
should be TPM normalized and log10 scaled.}

\item{signature_matrix}{a m x l matrix with m genes and l cell types. The
matrix should contain only a subset of the genes useful for the analysis.}

\item{...}{passed to the original seqImmuCC_LLSR fnuction}
}
\description{
Deconvolute using seqImmuCC (LLSR regression) and a custom signature matrix
}
