% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolute_timer}
\alias{deconvolute_timer}
\title{Deconvolute using the TIMER technique}
\usage{
deconvolute_timer(gene_expression_matrix, indications, ...)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples}

\item{indications}{a n-vector giving and indication string (e.g. 'brca') for each sample.}
}
\description{
Unlike the other methods, TIMER needs the specification of the
cancer type for each sample.
}
