% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolute_xcell}
\alias{deconvolute_xcell}
\title{Deconvolute using xCell}
\usage{
deconvolute_xcell(gene_expression_matrix, arrays,
  expected_cell_types = NULL, ...)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples}

\item{arrays}{Set to TRUE if microarray data, to FALSE for RNASeq (\code{rnaseq} parameter in xCell)}

\item{expected_cell_types}{a character list of the cell types to use
in the analysis. If NULL runs xCell with all cell types.
The spillover compensation step may over compensate, thus it is always better
to run xCell with a list of cell types that are expected to be in the mixture.
The names of cell types in this list must be a subset of the cell types that are inferred by xCell.
(\code{cell.types.use} parameter in xCell)}

\item{...}{Passed through to original xCell function. A native argument takes precedence
over an immunedeconv argument (e.g. \code{rnaseq} takes precedence over \code{arrays})
See \href{https://rdrr.io/github/dviraran/xCell/man/xCellAnalysis.html}{xCellAnalysis}}
}
\description{
Deconvolute using xCell
}
