% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolution_methods}
\alias{deconvolution_methods}
\title{List of supported immune deconvolution methods}
\usage{
deconvolution_methods
}
\description{
The methods currently supported are
\code{mcp_counter}, \code{epic}, \code{quantiseq}, \code{xcell}, \code{cibersort}, \code{cibersort_abs}, \code{timer}, \code{abis}, \code{consensus_tme}, \code{estimate}
}
\details{
The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.
}
