% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{.get_all_children}
\alias{.get_all_children}
\title{Recursive helper function for get_all_children.}
\usage{
.get_all_children(node, method)
}
\arguments{
\item{node}{Node in the cell type hierarchy. Will look for children of this node.}

\item{method}{method or dataset from the cell type mapping.}
}
\description{
Recursive helper function for get_all_children.
}
