% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{assign the following global variables
.onLoad because of 'staged-install'}
\usage{
.onLoad(libname, pkgname)
}
\description{
See https://developer.r-project.org/Blog/public/2019/02/14/staged-install/index.html
for more details.
}
