% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\docType{data}
\name{eig.sig}
\alias{eig.sig}
\title{Signature for eigengene}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 784 rows and 3 columns.
}
\usage{
eig.sig
}
\description{
Signature for eigengene
}
\keyword{datasets}
