% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{eset_to_matrix}
\alias{eset_to_matrix}
\title{Annotate unified cell_type names}
\usage{
eset_to_matrix(eset, column)
}
\arguments{
\item{eset}{\code{ExpressionSet}}

\item{column}{column name of the \code{fData()} table, which contains the HGNC gene symbols.}
}
\value{
matrix with gene symbols as rownames and sample identifiers as colnames.
}
\description{
(map the cell_types of the different methods to a common name)
convert an \code{Biobase::ExpressionSet} to a gene-expression matrix.
}
