% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_gene_expression_matrix}
\alias{example_gene_expression_matrix}
\title{example gene expression dataset}
\format{a gene expression matrix with gene_symbols as rownames and sample identifiers as colnames.

TODO describe how it was generated from fantom5 data or use a better dataset instead}
\usage{
example_gene_expression_matrix
}
\description{
example gene expression dataset
}
\keyword{datasets}
