% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{export_for_timer}
\alias{export_for_timer}
\title{Export gene expression matrix in a format that can be read by TIMER}
\usage{
export_for_timer(gene_expression_matrix, path = stop("Specify output path. "))
}
\arguments{
\item{gene_expression_matrix}{A numeric matrix with HGNC gene symbols as rownames and sample identifiers as colnames.}

\item{path}{path to write the TIMER-compatible tsv-file to.}
}
\description{
TIMER is only available as a web-tool. With this function,
a gene expression matrix can be exported in a format suitable for timer
with the same input gene expression matrix that can be used for the other methods.
}
