% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{find_children}
\alias{find_children}
\title{recursively traverse the cell type tree to resolve the mapping.}
\usage{
find_children(node, fractions, method)
}
\arguments{
\item{node}{data.tree::Node corresponding to a controlled vocabulary cell type}

\item{fractions}{a named list of fractions for each method_cell_type}

\item{method}{character identifing the method in the celltype_mapping}
}
\value{
numeric Either (1) the value of the method_cell_type mapped to cell_type,
(2) the sum of all child nodes (recursively) of cell_type
(3) NA, if the mapping cannot be resolved, i.e. at least one of the child nodes is missing.
}
\description{
recursively traverse the cell type tree to resolve the mapping.
}
