% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{get_all_children}
\alias{get_all_children}
\title{Get all children of a certain cell type.}
\usage{
get_all_children(cell_type, method = NULL)
}
\arguments{
\item{cell_type}{cell type name that appears in the cell type hierarchy.}

\item{method}{method or dataset from the cell type mapping.}
}
\value{
character vector of cell type names.
}
\description{
If method is NULL, this function will return ALL children of the given
cell type in the cell type hierarchy.
}
\details{
If a method is given, this function will return the cell type names
as used by \code{method} and stop as soon as a cell type maps to a cell type
provided by \code{method}.
}
