% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_cell_type_mapping.R
\name{get_cell_types_list}
\alias{get_cell_types_list}
\title{This function returns the list of all cell types in BASE/DCQ results,
along with the cell type they are mapped to}
\usage{
get_cell_types_list(method = c("dcq", "base"))
}
\arguments{
\item{method}{the method for which we want to get the cell types}
}
\value{
a dataframe with two columns: the exact cell type and the
correspondent, more general one
}
\description{
This function returns the list of all cell types in BASE/DCQ results,
along with the cell type they are mapped to
}
