% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_deconvolution_methods.R
\name{get_immucellai_mouse_layers}
\alias{get_immucellai_mouse_layers}
\title{List of cells in the different immucellai layers (mouse method)}
\usage{
get_immucellai_mouse_layers(layer = c("1", "2", "3"), print = FALSE)
}
\arguments{
\item{layer}{one of '1', '2', '3'}

\item{print}{logical, wether to print the cell types included in the layer}
}
\description{
List of cells in the different immucellai layers (mouse method)
}
