% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{GetFractions.Abbas}
\alias{GetFractions.Abbas}
\title{Constrained regression method implemented in Abbas et al., 2009}
\usage{
GetFractions.Abbas(XX, YY, w = NA)
}
\description{
Constrained regression method implemented in Abbas et al., 2009
}
