% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{import_from_timer}
\alias{import_from_timer}
\title{Import TIMER result}
\usage{
import_from_timer(input_file)
}
\arguments{
\item{input_file}{path to the input file.}
}
\description{
With this function a result table downloaded from timer an be imported and converted
to the unified result format that the other methods are using.
}
