% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{map_cell_type_to_controlled}
\alias{map_cell_type_to_controlled}
\title{Converts method-specific cell types to controlled vocabulary equivalents}
\usage{
map_cell_type_to_controlled(cell_type_list, method)
}
\arguments{
\item{cell_type_list}{vector of cell types we want CV for. If not found in
cell_type_map, returns NULL for that element}

\item{method}{method we are converting from}
}
\value{
vector with CV cell types as names
}
\description{
Converts method-specific cell types to controlled vocabulary equivalents
}
