% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{map_cell_types}
\alias{map_cell_types}
\title{Use a tree-hierarchy to map cell types among different methods.}
\usage{
map_cell_types(use_cell_types, fractions, method)
}
\arguments{
\item{use_cell_types}{list of cell types from the CV to map to}

\item{fractions}{named vector or list of cell type fractions. Names corresponds to the method_cell_types.}

\item{method1}{method or dataset with which \code{fractions} was generated}
}
\value{
numeric vector with CV cell types as names
}
\description{
\code{cell_type`` refers to a cell type from the controlled vocabulary (CV).}method_cell_type` refers to a cell type from a method or dataset.
}
