% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_datasets.R
\name{map_results_to_dataset}
\alias{map_results_to_dataset}
\title{annotate cell types with the cell types available in the dataset}
\usage{
map_results_to_dataset(results, which_dataset = available_datasets)
}
\description{
annotate cell types with the cell types available in the dataset
}
