% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_cell_type_mapping.R
\name{mouse_cell_type_mapping
Since DCQ and BASE provide estimates for several cell types, this function 
   combines the results to align them with the rest of the methods}
\alias{mouse_cell_type_mapping
Since DCQ and BASE provide estimates for several cell types, this function 
   combines the results to align them with the rest of the methods}
\alias{reduce_mouse_cell_types}
\title{Functions for mapping cell types between deconvolution methods.}
\usage{
reduce_mouse_cell_types(deconvolution.results, method = c("sum", "median"))
}
\arguments{
\item{deconvolution.results}{a dataframe containing the results of the
deconvolution experiment (DCQ or BASE). Cell types are in columns, samples in rows.}

\item{method}{the method used to combine the estimates: sum, for regression-based
methods as DCQ, or median for the scoring methods such as BASE.}
}
\value{
a data frame with the aggregated cell types
}
\description{
Functions for mapping cell types between deconvolution methods.
}
