% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_cell_type_mapping.R
\name{mouse_cell_type_mapping}
\alias{mouse_cell_type_mapping}
\title{Functions for mapping cell types between deconvolution methods.}
\description{
Functions for mapping cell types between deconvolution methods.
}
