% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_deconvolution_methods.R
\name{mouse_genes_to_human}
\alias{mouse_genes_to_human}
\title{This function converts the mouse gene symbols into corresponding human ones.}
\usage{
mouse_genes_to_human(gene.expression.matrix)
}
\arguments{
\item{gene.expression.matrix}{a m x n matrix with m genes and n samples.
Gene symbols must be the rownames of the matrix.}
}
\value{
the same matrix, with the counts for the corresponding human genes.
This matrix can directly be used with the immunedeconv methods. A message
will display the ratio of original genes which were converted.
}
\description{
This function converts the mouse gene symbols into corresponding human ones.
}
