% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{node_by_name}
\alias{node_by_name}
\title{Lookup dictionary for cell-type nodes}
\usage{
node_by_name
}
\description{
Access nodes by name in O(1). Node names are unique in our tree.
gets attached on .onLoad, see zzz.R
}
