% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb.R
\name{perturb_sig}
\alias{perturb_sig}
\title{perturb_sig}
\usage{
perturb_sig(sig, prop_gene_rm, input_melted = TRUE, melt_output = FALSE)
}
\arguments{
\item{sig}{data.frame. If input_melted is true, should be:
where first column is gene,
second column is cell type, third columns value for that gene,type pair
third column is optional.
If input_melted is false, rownames are genes, columns are cell types}

\item{melt_output}{boolean whether to return the signature matrix in
melted format or not. If \code{FALSE}, returns n_gene x n_cell_type data.frame}

\item{n_gene_rm}{number of genes per cell type to randomly remove}

\item{melt_input}{boolean whether input sig is melted or not}
}
\description{
perturb_sig
}
