% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_cell_type_mapping.R
\name{reduce_mouse_cell_types}
\alias{reduce_mouse_cell_types}
\title{Since DCQ and BASE provide estimates for several cell types, this function
combines the results to align them with the rest of the methods}
\usage{
reduce_mouse_cell_types(deconvolution.results, method = c("sum", "median"))
}
\arguments{
\item{deconvolution.results}{a dataframe containing the results of the
deconvolution experiment (DCQ or BASE). Cell types are in columns, samples in rows.}

\item{method}{the method used to combine the estimates: sum, for regression-based
methods as DCQ, or median for the scoring methods such as BASE.}
}
\value{
a data frame with the aggregated cell types
}
\description{
Since DCQ and BASE provide estimates for several cell types, this function
combines the results to align them with the rest of the methods
}
