% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_simulation.R
\name{scale_to_million}
\alias{scale_to_million}
\title{Scale sample to TPM}
\usage{
scale_to_million(sample)
}
\arguments{
\item{sample}{numeric vector of gene expression values}
}
\description{
a TPM sample always sums to 1,000,000.
}
\details{
This function takes an input vector and
rescales it, so that the sum of the vector equals 1e6.
}
