% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqImmuCC_LLSR.R
\name{seqImmuCC_LLSR}
\alias{seqImmuCC_LLSR}
\title{Source code for the Linear Least Square Regression performed in seqImmuCC.}
\usage{
seqImmuCC_LLSR(
  signature,
  SampleData,
  w = NA,
  QN = T,
  sig.scale = F,
  sig.stand = T,
  sample.scale = T,
  log = T
)
}
\arguments{
\item{signature}{signature matrix}

\item{SampleData}{sample expression profile}

\item{w}{vector to perform weighted least squares}

\item{QN}{logical. If TRUE expression matrix is quantile normalized}

\item{sig.scale}{logical. If TRUE, scales the signature matrix}

\item{sig.stand}{logical.If TRUE, standardizes the signature matrix}

\item{sample.scale}{logical. If TRUE, scales the sample expression}

\item{log}{logical. If TRUE, log transforms signature and expression data}
}
\description{
This code is provided by the authors and the method available at:
http://218.4.234.74:3200/immune/
}
\details{
The method is described in Chen et al. doi: 10.3389/fimmu.2018.01286
}
