% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{set_cibersort_mat}
\alias{set_cibersort_mat}
\title{Set Path to CIBERSORT matrix file (\code{LM22.txt})}
\usage{
set_cibersort_mat(path)
}
\description{
CIBERSORT is only freely available to academic users.
A license an the binary can be obtained from https://cibersort.stanford.edu.
}
