% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_simulation.R
\name{single_cell_simulation}
\alias{single_cell_simulation}
\title{Functions for generating simulated bulk samples from a single cell dataset.}
\description{
Functions for generating simulated bulk samples from a single cell dataset.
}
