% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{summarise_children}
\alias{summarise_children}
\title{sum up the fraction of all child nodes to a single value.
Take into account which cell types are optional.}
\usage{
summarise_children(df)
}
\arguments{
\item{df}{with samples in columns and cell types in rows. Contains a logical columnn \code{optional} which states which
cell type may be missing}
}
\description{
sum up the fraction of all child nodes to a single value.
Take into account which cell types are optional.
}
