% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer_available_cancers}
\alias{timer_available_cancers}
\title{TIMER signatures are cancer specific. This is the list of available cancer types.}
\usage{
timer_available_cancers
}
\description{
TIMER signatures are cancer specific. This is the list of available cancer types.
}
