% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb.R
\name{truncate_signature}
\alias{truncate_signature}
\title{truncate_signature}
\usage{
truncate_signature(signature, n_gene_rm, return_melted = TRUE)
}
\arguments{
\item{signature}{data.frame where first column is gene,
second column is cell type, third columns value for that gene,type pair
third column is optional}

\item{n_gene_rm}{number of genes per cell type to randomly remove}

\item{return_melted}{boolean whether to return the signature matrix in
melted format or not. If \code{FALSE}, returns n_gene x n_cell_type data.frame}
}
\description{
truncate_signature
}
