% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{xCell.data}
\alias{xCell.data}
\title{Data object from xCell.}
\usage{
xCell.data
}
\description{
For some reason, this object is not properly exported from the xCell namespace.
This is a workaround, that \code{xCellAnalysis} can be properly called from this package.
}
