% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_datasets.R
\docType{data}
\name{available_datasets}
\alias{available_datasets}
\title{List of available validation datasets}
\format{An object of class \code{character} of length 5.}
\usage{
available_datasets
}
\description{
List of available validation datasets
}
\keyword{datasets}
