% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\docType{data}
\name{available_datasets}
\alias{available_datasets}
\title{Available methods and datasets.}
\format{An object of class \code{character} of length 12.}
\usage{
available_datasets
}
\description{
A list of all methods (e.g. \code{cibersort}) and datasets (e.g. \code{schelker_ovarian}) for
which the cell types are mapped to the controlled vocabulary.
}
\keyword{datasets}
