% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{cell_type_map}
\alias{cell_type_map}
\title{Table mapping the cell types from methods/datasets to a single, controlled vocabulary.}
\description{
Columns: \code{method_dataset}, \code{method_cell_type}, \code{cell_type}.
}
\details{
See \code{inst/extdata/cell_type_mapping.xlsx} for more details.
}
