% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\docType{data}
\name{cell_type_map}
\alias{cell_type_map}
\title{Table mapping the cell types from methods/datasets to a single, controlled vocabulary.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 133 rows and 3 columns.}
\usage{
cell_type_map
}
\description{
Columns: \code{method_dataset}, \code{method_cell_type}, \code{cell_type}.
}
\details{
See \code{inst/extdata/cell_type_mapping.xlsx} for more details.
}
\keyword{datasets}
