% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{cell_type_mapping}
\alias{cell_type_mapping}
\title{Functions for mapping cell types between deconvolution methods and datasets.}
\description{
makes use of the manually curated mapping in \code{inst/extdata/cell_type_mapping.xlsx}.
}
\details{
TEST TEST TODO
}
