% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\docType{data}
\name{cell_type_mapping}
\alias{cell_type_mapping}
\title{Functions for mapping cell types between deconvolution methods and datasets.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
cell_type_mapping
}
\description{
makes use of the manually curated mapping in \code{inst/extdata/cell_type_mapping.xlsx}.
}
\keyword{datasets}
