% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\docType{data}
\name{cell_type_tree}
\alias{cell_type_tree}
\title{Available cell types in the controlled vocabulary organized as a lineage tree.}
\format{
An object of class \code{Node} (inherits from \code{R6}) of length 42.
}
\usage{
cell_type_tree
}
\description{
Available cell types in the controlled vocabulary organized as a lineage tree.
}
\details{
a \code{data.tree} object
}
\keyword{datasets}
