% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{check_cancer_types}
\alias{check_cancer_types}
\title{process batch table and check cancer types.}
\usage{
check_cancer_types(args)
}
\arguments{
\item{args}{environment}
}
\description{
process batch table and check cancer types.
}
