% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_deconvolution_methods.R
\docType{data}
\name{custom_deconvolution_methods}
\alias{custom_deconvolution_methods}
\title{List of methods that support the use of a custom signature}
\format{
An object of class \code{character} of length 4.
}
\usage{
custom_deconvolution_methods
}
\description{
The available methods are
\code{epic}, \code{cibersort}, \code{cibersort_abs}, \code{consensus_tme}, \code{seqimmucc}
}
\details{
The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.
}
\keyword{datasets}
