% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_petitprez}
\alias{dataset_petitprez}
\title{RNA-seq dataset with FACS measurements}
\format{
TODO
}
\source{
TODO
}
\usage{
dataset_petitprez
}
\description{
TODO
}
\keyword{datasets}
