% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\name{deconvolute}
\alias{deconvolute}
\title{Perform an immune cell deconvolution on a dataset.}
\usage{
deconvolute(
  gene_expression,
  method = deconvolution_methods,
  indications = NULL,
  tumor = TRUE,
  arrays = FALSE,
  column = "gene_symbol",
  rmgenes = NULL,
  scale_mrna = TRUE,
  expected_cell_types = NULL,
  ...
)
}
\arguments{
\item{gene_expression}{A gene expression matrix or a Biobase ExpressionSet.
Either: A numeric matrix or data.frame with HGNC gene symbols as rownames and sample identifiers as colnames.
Or: A Biobase ExpressionSet with HGNC symbols in an fData column (see \code{column} parameter)
In both cases, data must be on non-log scale.}

\item{method}{a string specifying the method.
Supported methods are \code{xcell}, \code{...}}

\item{indications}{a character vector with one indication per
sample for TIMER. Argument is ignored for all other methods.}

\item{tumor}{use a signature matrix/procedure optimized for tumor samples,
if supported by the method. Currently affects EPIC and quanTIseq.}

\item{arrays}{Runs methods in a mode optimized for microarray data.
Currently affects quanTIseq and CIBERSORT.}

\item{column}{Only relevant if \code{gene_expression} is an ExpressionSet. Defines in which column
of fData the HGNC symbol can be found.}

\item{rmgenes}{a character vector of gene symbols. Exclude these genes from the analysis.
Use this to exclude e.g. noisy genes.}

\item{scale_mrna}{logical. If FALSE, disable correction for mRNA content of different cell types.
This is supported by methods that compute an absolute score (EPIC and quanTIseq)}

\item{expected_cell_types}{Limit the analysis to the cell types given in this list. If the cell
types present in the sample are known \emph{a priori}, setting this can improve results for
xCell (see https://github.com/grst/immunedeconv/issues/1).}

\item{...}{arguments passed to the respective method}
}
\value{
\code{data.frame} with \code{cell_type} as first column and a column with the
calculated cell fractions for each sample.
}
\description{
Perform an immune cell deconvolution on a dataset.
}
\examples{
# Not run: deconvolute(gene_expression_matrix, "xcell")

}
