% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\docType{methods}
\name{deconvolute}
\alias{deconvolute}
\alias{deconvolute.default}
\alias{deconvolute}
\alias{deconvolute,eSet,character-method}
\alias{deconvolute,matrix,character-method}
\title{Add together two numbers.}
\usage{
deconvolute.default(gene_expression, method = deconvolution_methods, ...)

deconvolute(gene_expression, method = deconvolution_methods,
  column = "gene_symbol")

\S4method{deconvolute}{eSet,character}(gene_expression,
  method = deconvolution_methods, column = "gene_symbol")

\S4method{deconvolute}{matrix,character}(gene_expression, method)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{eset}{Expression set}

\item{col}{column name of the \code{fData} column that contains HGNC gene symbols.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\section{Methods (by class)}{
\itemize{
\item \code{gene_expression = eSet,method = character}: \code{eset} is a \code{Biobase::ExpressionSet}.
\code{fData} contains a column with HGNC gene symbols (specify column name)

\item \code{gene_expression = matrix,method = character}: \code{matrix} is matrix with HGNC gene symbols as row names.
}}

\examples{
add(1, 1)
add(10, 1)
}
