% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\docType{methods}
\name{deconvolute}
\alias{deconvolute}
\alias{deconvolute.default}
\alias{deconvolute}
\alias{deconvolute,eSet,character,character,character-method}
\alias{deconvolute,eSet,character,character,missing-method}
\alias{deconvolute,matrix,character,missing,character-method}
\alias{deconvolute,matrix,character,missing,missing-method}
\alias{deconvolute,data.frame,character,missing,character-method}
\alias{deconvolute,data.frame,character,missing,missing-method}
\title{Perform an immune cell deconvolution on a dataset.}
\usage{
deconvolute.default(gene_expression, method = deconvolution_methods,
  indications = NULL, tumor = TRUE, arrays = FALSE)

deconvolute(gene_expression, method = deconvolution_methods,
  column = "gene_symbol", indications = NULL)

\S4method{deconvolute}{eSet,character,character,character}(gene_expression,
  method = deconvolution_methods, column = "gene_symbol",
  indications = NULL)

\S4method{deconvolute}{eSet,character,character,missing}(gene_expression,
  method, column = "gene_symbol")

\S4method{deconvolute}{matrix,character,missing,character}(gene_expression,
  method, indications = NULL)

\S4method{deconvolute}{matrix,character,missing,missing}(gene_expression,
  method)

\S4method{deconvolute}{data.frame,character,missing,character}(gene_expression,
  method, indications = NULL)

\S4method{deconvolute}{data.frame,character,missing,missing}(gene_expression,
  method)
}
\arguments{
\item{gene_expression}{A numeric matrix with HGNC gene symbols as rownames and sample identifiers as colnames.
Data must be on non-log scale.}

\item{method}{a string specifying the method.
Supported methods are \code{xcell}, \code{...}}

\item{indications}{a character vector with one indication per
sample for TIMER. Argument is ignored for all other methods.}

\item{tumor}{use a signature matrix/procedure optimized for tumor samples,
if supported by the method. Currently affects EPIC and quanTIseq.}

\item{arrays}{Runs methods in a mode optimized for microarray data.
Currently affects quanTIseq and CIBERSORT.}

\item{...}{arguments passed to the respective method}

\item{eset}{Expression set}

\item{col}{column name of the \code{fData} column that contains HGNC gene symbols.}
}
\value{
\code{data.frame} with \code{cell_type} as first column and a column with the
calculated cell fractions for each sample.
}
\description{
Perform an immune cell deconvolution on a dataset.
}
\section{Methods (by class)}{
\itemize{
\item \code{gene_expression = eSet,method = character,column = character,indications = character}: \code{eset} is a \code{Biobase::ExpressionSet}.
\code{fData} contains a column with HGNC gene symbols (specify column name)

\item \code{gene_expression = eSet,method = character,column = character,indications = missing}: variant where indications are not defined, cannot be used with method == 'timer'

\item \code{gene_expression = matrix,method = character,column = missing,indications = character}: \code{matrix} is matrix with HGNC gene symbols as row names.

\item \code{gene_expression = matrix,method = character,column = missing,indications = missing}: variant where indications are not defined, cannot be used with method=='timer'

\item \code{gene_expression = data.frame,method = character,column = missing,indications = character}: \code{matrix} is data frame with HGNC gene symbols as row names.

\item \code{gene_expression = data.frame,method = character,column = missing,indications = missing}: variant where indications are not defined, cannot be used with method=='timer'
}}

\examples{
# Not run: deconvolute(gene_expression_matrix, "xcell")

}
