% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_deconvolution_methods.R
\name{deconvolute_dcq}
\alias{deconvolute_dcq}
\title{Deconvolute using DCQ}
\usage{
deconvolute_dcq(
  gene.expression.matrix,
  ref.samples = NULL,
  n.repeats = 10,
  combine_cells = TRUE,
  ...
)
}
\arguments{
\item{gene.expression.matrix}{a m x n matrix with m genes and n samples}

\item{ref.samples}{a vector containing the names of the samples to be
used to standardize the expressions. If NULL, all the samples are used.}

\item{n.repeats}{the number of models to generate. Predicted cell quantities
will then be averaged.}

\item{combine_cells}{logical. This method estimates several cell types (~200).
If TRUE (default), these are combined into 19 major immune-stromal cell types.}
}
\description{
Deconvolute using DCQ
}
