% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_deconvolution_methods.R
\name{deconvolute_mouse}
\alias{deconvolute_mouse}
\title{Perform deconvolution on a mouse RNAseq dataset}
\usage{
deconvolute_mouse(
  gene_expression_matrix,
  method = deconvolution_methods_mouse,
  rmgenes = NULL,
  column = "gene_symbol",
  algorithm = NULL,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{a m x n matrix with m genes and n samples.
Data should be TPM normalized, except when using seqImmuCC where raw
counts are prefereable.}

\item{method}{string specifying the method}

\item{rmgenes}{noisy genes to be removed from the analysis}

\item{...}{additional arguments passed to the respective method}
}
\value{
a dataset with the estimated fractions/scores, with samples in rows and cell types in column
}
\description{
Perform deconvolution on a mouse RNAseq dataset
}
