% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\docType{data}
\name{deconvolution_methods}
\alias{deconvolution_methods}
\title{list of supported immune deconvolution methods}
\format{An object of class \code{character} of length 4.}
\usage{
deconvolution_methods
}
\description{
list of supported immune deconvolution methods
}
\keyword{datasets}
