% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\docType{data}
\name{deconvolution_methods}
\alias{deconvolution_methods}
\title{List of supported immune deconvolution methods}
\format{
An object of class \code{character} of length 7.
}
\usage{
deconvolution_methods
}
\description{
The methods currently supported are
\code{mcp_counter}, \code{epic}, \code{quantiseq}, \code{xcell}, \code{cibersort}, \code{cibersort_abs}, \code{timer}
}
\details{
The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.
}
\keyword{datasets}
