% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_deconvolution_methods.R,
%   R/immune_deconvolution_methods.R
\docType{data}
\name{deconvolution_methods}
\alias{deconvolution_methods}
\title{List of methods that support the use of a custom signature}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 9.
}
\usage{
deconvolution_methods

deconvolution_methods
}
\description{
The available methods are
\code{epic}, \code{cibersort}, \code{cibersort_abs}, \code{consensus_tme}, \code{base}

The methods currently supported are
\code{mcp_counter}, \code{epic}, \code{quantiseq}, \code{xcell}, \code{cibersort}, \code{cibersort_abs}, \code{timer}, \code{abis}, \code{consensus_tme}
}
\details{
The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.

The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.
}
\keyword{datasets}
