% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_deconvolution_methods.R
\docType{data}
\name{deconvolution_methods_mouse}
\alias{deconvolution_methods_mouse}
\title{List of supported mouse deconvolution methods
The methods currently supported are
\code{mmcp_counter}, \code{seqimmucc}, \code{dcq}, \code{base}
The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.}
\format{
An object of class \code{character} of length 4.
}
\usage{
deconvolution_methods_mouse
}
\description{
List of supported mouse deconvolution methods
The methods currently supported are
\code{mmcp_counter}, \code{seqimmucc}, \code{dcq}, \code{base}
The object is a named vector. The names correspond to the display name of the method,
the values to the internal name.
}
\keyword{datasets}
