% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_mapping.R
\name{find_children}
\alias{find_children}
\title{Recursive helper function for map_cell_types.}
\usage{
find_children(node, fractions, method_dataset = NULL)
}
\arguments{
\item{node}{data.tree::Node corresponding to a controlled vocabulary cell type}

\item{fractions}{a named list of fractions for each method_cell_type}

\item{method_dataset}{character identifying the method or dataset in the celltype_mapping. If NULL, the input
vector already uses controlled vocabulary.}
}
\value{
numeric Either (1) the value of the method_cell_type mapped to cell_type,
(2) the sum of all child nodes (recursively) of cell_type
(3) NA, if the mapping cannot be resolved, i.e. at least one of the child nodes is missing.
}
\description{
Traverses the cell type hierarchy, summing up cell type fractions when necessary.
}
