% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immune_deconvolution_methods.R
\docType{package}
\name{immunedeconv}
\alias{immunedeconv}
\alias{immunedeconv-package}
\title{Collection of immune cell deconvolution methods.}
\description{
Collection of immune cell deconvolution methods.
}
