% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_simulation.R
\name{make_bulk_eset}
\alias{make_bulk_eset}
\title{Make a random expression set from a single-cell dataset}
\usage{
make_bulk_eset(eset, cell_fractions, n_cells = 500, combine = mean)
}
\arguments{
\item{eset}{\code{Biobase::ExpressionSet} with a \code{cell_type} column in \code{pData}.}

\item{cell_fractions}{n x n_cell_types dataframe with the fraction for each
sample in each row.}

\item{n_cells}{number of single cells to use in each sample}

\item{combine}{callback function used to aggregate the counts.}
}
\description{
Make a random expression set from a single-cell dataset
}
\seealso{
make_random_bulk
}
