% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_datasets.R
\name{make_bulk_eset}
\alias{make_bulk_eset}
\title{make a random expression set from a single-cell dataset}
\usage{
make_bulk_eset(eset, cell_fractions, n_cells = 500, combine = sum)
}
\arguments{
\item{eset}{\code{Biobase::ExpressionSet} with a \code{cell_type} column in \code{pData}.}

\item{cell_fractions}{n x n_cell_types dataframe with the fraction for each
sample in each row.}

\item{combine}{callback function used to aggregate the counts.}

\item{n_cell}{number of single cells to use in each sample}
}
\description{
make a random expression set from a single-cell dataset
}
