% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_datasets.R
\name{make_random_bulk}
\alias{make_random_bulk}
\title{make a random bulk sample from a single cell dataset}
\usage{
make_random_bulk(eset, n_cells = 500)
}
\description{
make a random bulk sample from a single cell dataset
}
