% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell-type-mapping.R
\name{map_cell_types}
\alias{map_cell_types}
\title{Use a tree-hierarchy to map cell types among different methods.}
\usage{
map_cell_types(use_cell_types, fractions1, fractions2, method1, method2)
}
\arguments{
\item{use_cell_types}{list of cell types from the controlled vocabulary to map}

\item{fractions1}{named vector or list of cell type fractions. Name corresponds to the cell types}

\item{fractions2}{}

\item{method1}{method or dataset to use for mapping \code{fractions1}}

\item{method2}{}
}
\value{
named vector. Name of the vector corresponds to the controlled vocabulary cell types.
}
\description{
Use a tree-hierarchy to map cell types among different methods.
}
