% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_simulation.R
\name{scale_to_million}
\alias{scale_to_million}
\title{scale sample to TPM}
\usage{
scale_to_million(sample)
}
\arguments{
\item{sample}{numeric vector of gene expression values}
}
\description{
an TPM sample always sums to 1,000,000.
}
