% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\docType{data}
\name{timer_available_cancers}
\alias{timer_available_cancers}
\title{TIMER signatures are cancer specific. This is the list of available cancer types.}
\format{
An object of class \code{character} of length 32.
}
\usage{
timer_available_cancers
}
\description{
TIMER signatures are cancer specific. This is the list of available cancer types.
}
\keyword{datasets}
